Implement the chart selection mega-extenda-bar
On selection in one chart, pmchart now reflects that
selection along the time axis (X) into all of the other
charts.  Implemented by simulation of mouse events in
those other charts, in a cunning way that engages the
picker state machine but does not perform a real actual
selection in the other charts.