Fix the python API wrappers for reentrant libpcp string routines
The python wrappers for pmErrStr, pmIDStr, pmiErrStr, pmUnitsStr,
pmInDomStr, pmTypeStr, and pmAtomStr were not correct - they were
using the reentrant C interfaces but not passing in a buffer (nor
a size).  Imploded on ppc64.

Adds in missing python dictionary symbols for maximum err string
lengths, and uses these for the two error interfaces.