qa/src - problem with ../localconfig
This commit fixes the "test: 1: -ge: unexpected operator" error in
the build.  The root cause was a hangover from merging the separate
QA tree into the pcp tree ... localconfig is needed by qa/src/GNUmakefile
but it is created by qa/GNUmakefile ... works for a build in place, but
not the src-link step used by Makepkgs.  Fixed with some gmake magic.