Avoid QAssistantClient on platforms where its not shipped
In current Mac OS X and Windows builds of Qt (4.7 onward)
there is no QAssistantClient class anymore.  On Linux the
distros build the back-compat class and ship it so we can
continue to use it there.

Longer term, we need to switch the help system away from
this class to something browser-based.