Fix pmstat handling of odd corner case in CPU metrics
Came across a machine which was exhibiting a kernel bug,
possibly in the Radeon driver, such that all CPU time was
being consumed in the kernel.  Its likely interrupts were
disabled for extended periods too, as on a 1sec sampling
interval, pmstat would often observe no change in the CPU
time metrics.

When this unusual situation arises, pmstat attempts to
divide by zero and takes sigfpe - guard against this, and
handle it as if no values were observed (which is fairly
close to true).