Fix intermittent failure of test qa/131.
If the initial connection timed out (as expected) all was well.
However, sometimes the connection was successful and the handshake
would time out. In this case, pminfo reported an IPC protocol
failure. This was because __pmConnectHandshake did not
propagate the PM_ERR_TIMEOUT from __pmGetPDU to the caller and
instead was retuning PM_ERR_IPC for all errors except PDU_ERROR.

__pmConnectHandshake now propagates the PM_ERR_TIMEOUT. The test
script has been enhanced to handle the message which occurs when
this happens.