Fix issues with pmdashping timeouts observed on multi-core x86_64
Test qa/323 has been failing frequently for some time here.  Root
cause turned out to be signal related again, this time interacting
poorly with pthreads.  A race existed between the calls to signal()
in the main() thread vs the refresh() thread - in reality, signals
are process-wide and cannot be directed to specific threads.  So,
whether timeouts were applied at all or not depended on the order
of calls to signal(), which could go either way but often ended up
not being applied.