A more general solution for filtering dots in "Waiting for pmcd" messages
Filter strings of 3 or more dots in "Waiting for pmcd" messages in
filter_pcp_start and filter_pcp_stop rather than in the individual
tests. Replace strings of 3 or more dots with a string of 3 dots
rather than with the string '[dots]'. This fixes many individual
tests without editing them to expect '[dots]'