Change __pmSetProcessIdentity to fail more resoundingly
For good security reasons, if we cannot change to the intended
user identity we should abort operation in preference to simply
continuing on as we were.  All callers need to fail hard in this
scenario, so log a critical error and call exit.  This is a bit
like pmGetConfig in terms of level-of-trouble-that-you-are-in if
it ever happens, so exiting seems an acceptable option.

Adds a QA test (653) to exercise the good and bad cases.

Resolves PCP bugzilla bug #933.