Handle EINPROGRESS status from PR_Connect.
NSPR manages all file descriptors in nonblocking mode, so it
is possible to get PR_FAILURE from PR_Connect when connecting to
a socket with oserror() returning EINPROGRESS. In this case,
the client must wait for the socket using __pmSelect{Read,WRite}
and then call __pmConnectCheckError in order to determine
whether the connection was successful.

This fixes qa test case 439 for NSS/NSPR enabled builds.