Fix QA fallout from making __pmSetProcessIdentity exit(1)
Previous to this change, a number of QA tests were passing
because the user running QA was still able to exercise the
dbpmda/pmcd/pmlogger_check code as themselves.  This was as
a result of __pmSetProcessIdentity being permissive of any
failed attempt to change user (just logging a warning and
continuing on, which was fine for QA).

Now, we jump through hoops in order to either ensure these
tests are started by a user with sufficient privileges that
they can successfully change user, or to use the -U options
to override the default preference of changing to the "pcp"
user.