Fix race conditions in pmie and pmlogger startup scripts
Recent changes to tempfile handling has regressed the pmie and
pmlogger startup scripts.  Errors of the form:
/etc/rc.d/init.d/pmlogger: line 100: /var/tmp/pcp.5vfQsSHKo/pmcheck: No such file or directory
are now produced.

Because sections of these two scripts are run in parallel with
the original script, we are open to race conditions where the
main script exits and removes the temporary directory before the
co-process has an opportunity to create its temporary file.  We
can resolve this using separate temporary dirs and trap handling
which thus no longer race.

Worse still, QA failed to detect the problem.  At least one test
that should have found the problem trivially (575) failed to, as
a result of aggressively discarding stderr and stdout instead of
using filtering.  This has been rectified and common pmie filter
routines abstracted (from tests 115, 504 and 575) to simplify the
task for all current and future tests.