Add an interface allowing PMAPI clients to explicitly cleanup
Introduce a new __pmShutdown routine which wraps up all needed
teardown of resources allocated in libpcp.  It initially covers
the NSS cleanup and DSO PMDA end-callback / dlclose() for any
local context PMDAs the client opened.  Takes care of ordering
concerns in shutdown between different areas, and works nicely
with the existing atexit handler for local context teardown.

Conditionally uses it in qa/src, and valgrind is now content
with the workings of test 387 (with NSS enabled).