Implement support for pmcd file descriptor leak checker with NSS
Test 578 was failing with NSS, as a result of changes around the
way the file descriptor table indexing works, and the semantics
of the return value of __pmSocket type routines.

This change keeps the existing semantics after some discussion of
possible alternatives on the list, the alternative of tracking the
actual count of open file descriptors proved far too difficult in
practice.