Add an internal __pmGetAPIConfig feature-test interface
This implements a new feature testing interface suitable for
QA testing and any applications which might need to enquire
about support for features they intend to use.

The pmconfig(1) command is extended with a -L option allowing
one to call this from the command line, in both an "all" mode
and a "specific feature" query mode.