kvm PMDA needs to run as root
The kvm PMDA cannot change to user "pcp", because the directory
/sys/kernel/debug/kvm is unreadable by mere mortals.

Needs force_restart in the Install script ... SIGHUP to a pmcd
tha has already downgraded from "root" to "pcp" does not work.

And no $pmda->set_user('pcp') call (the latter was guarded by if
-r $kvm_path which is true when you're running as "root", but as
soon as the user is changed to "pcp" this is no longer true and
the PMDA returns no metric values).