Remove the unused -c option from pmdasummary
After some discussion, noone remembers what this was for,
everyone is sure the "Array Services" product mentioned on
the man page is pre-historical, and the usage message that
points to use of ASCII PDUs was the nail in the coffin.

Cannot possibly be in-use-by anyone, so have removed it
rather than attempt to explain it via docs.