Initial PCP protocol changes to support secure sockets layer (SSL)
This introduces several changes that prepare for the pmcd/client
protocol exchange to be optionally augmented with an SSL handshake
after the initial PCP handshake.

Enabling this requires several steps.  Firstly, the PM_CTXFLAG concept
is reintroduced (which was dropped awhile ago from lack of interest).
The shallow/exclusive flags that were there are left alone, for back-
compat, and a new SECURE flag is added for use with HOST contexts.  In
addition, in the future, pmParseHostSpec is likely to be extended with
pcp://... and pcps://... prefix extraction.  Current intention is to
wait on deeper understanding of user authentication before that gets
tackled, however, for a more coherent naming strategy.

When a socket is established to pmcd, the first stage of the protocol
is to send an extended error PDU back to the client.  This contains a
field that was previously always zeroed - this is now reclaimed to
hold "feature" flags.  If pmcd is capable of doing SSL encryption, a
bit is set here indicating the presense of the feature (at least 15
other bits are also available for later features, at which point we
could also go with a larger "pduinfo" payload, I think).

All existing client tools ignore this bit, and are oblivious to the
fact that it changed from a zero to a one for pmcd binaries built with
this feature.  For new clients who choose to take pmcd up on its offer
of encrypted communication, the credentials version PDU (which the
client sends back) has also been modified, in a similar way - it has a
previously-always-zero bit now set, which is used in pmcd to identify
client willingness to partake.

Early stages of the subsequent SSL exchange are in place, but there's
plenty of work to do in completing that (around setting up certificate
databases and so forth in pmcd).

A large part of this change is around reworking __pmDataIPC & friends
to be able to hold more than just a single pointer (it now also holds
an SSL fd, in addition to the existing NSPR).  This is also expected to
grow in future.  If NSS (--without-secure-sockets) is disables during
configure, the ipc table is now returned to its pre-NSS size too, an
unexpected little bonus.