Server side (pmcd) support for encrypted and compressed communication
Adds the necessary SSL support code for responding to a client request
for SSL connections.  Two new pmcd command line options are added, one
allowing a non-default credentials database path to be specified and
the other to allow a password file to be passed in if the database is
password protected.  At startup the database is queried for credential
nickname "PCP Collector credential", which is then verified, private
key extracted, and used for subsequent encrypted channels (both clear
and encrypted communication is supported by the one pmcd, just depends
on what an individual client requests for each context it creates).

Its not clear which of these new routines will end up in libpcp and
which in pmcd (some will be shared with client side), so for now all
have used libpcp naming conventions and I'll revisit that later when
cleaning up.