PCP_SYSCONF_DIR changes
Big changes here to introduce $PCP_SYSCONF_DIR (defaults to
/etc/pcp) as the base directory for the configuration files for
pmcd, pmie, pmlogger and pmproxy.

Also included is the migration script (debian/cleanconfigs)
that can be used in packaging as part of the pre-install script
to harvest and migrate configuration files from all the places
we've hidden them in the past.