pmie - fix corner case fetch profile handling bug
Found during QA/514 development.

If we have multiple rules, with metrics (or the same metric) over a
common instance domain, and _some_ rules include explicit instance
names (e.g. #foo #'foo-bar') and _some_ rules have no instance
names (implying all instances), then we _may_ botch the handling
of the fetch profile and fetch fewer instances than are required.

The effect depends on the order of the rules in the configuration
file and the non-deterministic order of rule evaluation after
connection to pmcd is made (or re-established).  So, this is a
little ellusive!

Fortunately the pre-conditions are probably rarely seen in real
production rule sets.