pmie - retire impossible count scale factors
The count scale factors (keywords) Gcount (10^9) and Tcount
(10^12) cannot possibly produce the correct results, as their
values (9 and 12 respectively) are stuffed into the scaleCount
field of a pmUnits struct ... but scaleCount is a 4-bit _signed_
int, so the maximum positive value it can take is 7.

Quietly retire Gcount and Tcount from the lexical scanner and the
man page ... oops.

Found while developing QA 521.