Abstract the code dealing with errno-based socket close detection
Now that we have a mix of NSPR and regular errno values, we need a
different test for socket closure.  Turn it into a switch, and add
the NSPR codes into the mix.  Resolves an issue where pmcd logged
an error on every client connect, due to NSPR codes now being more
widely exposed.