Final commit allowing SSL connections to function correctly.
This commit resolves the two remaining issues surrounding SSL
connections (that I know of, and exercised by QA test 712).
Firstly, the connection was hanging at various points in the
client side following a successful initial handshake.  Turns
out the SSL code can consume data beyond the initial part of
a PDU read, resulting in a subsequent select() prior to the
end of the PDU hanging (due to all data being transparently
consumed previously).  After some searching, an SSL interface
to test for this condition was found (SSL_DataPending), which
must be passed an SSL fd (else segv results, evidently). This
is a bit of a pain, as it means we do need to distinguish the
NSPR/SSL descriptors after all (or at least flag a successful
ssl fd overwrite).

Secondly, the server code was taking sigpipe at the end of a
successful client SSL interaction, as a result of a redundant
(duplicate) PR_Close call on an already-closed descriptor.

Test 712 now passes reliably.  Happy days!