Initialize the supplementary group access list for all daemons
Ensure __pmSetProcessIdentity correctly initialises supplementary
group access list, so that we don't end up with whatever was set
there by the caller (which can include group 0 and other system
groups).

Update pmproxy to call __pmSetProcessIdentity instead of the open
coded implementation it had previously (from the days when it ran
as user "nobody").