Cleanup recently introduced SSL_DataPending check wrapper
Initially attempted to get a direct call to PR_Poll working
for this (with no SSL-specific code), but cannot seem to make
that work with our native-select native-and-nspr-fd handling
methods.  So, fell back to removing __pmSecureDataPending and
making a more appropriate/generic wrapper which allows for a
SSL test in secure-sockets mode only.