libpcp - fix alignment issues
We were being a little loose with pointer assignments involving
pmAtomValue types and values stored in the vbuf[] field of a
pmValueBlock.

On ia64 this was generating a lot of alignment warnings at
run time when memcpy() was being in-lined by gcc.

Fixed by being more careful and more C99 standards-compliant.