Improve pmdaapache fetch handling
Ryan reported and fixed a pmdaapache fd leak.  He also noticed
that one descriptor per metric was leaked on each fetch, which
suggested the http request was being performed too often.

On closer inspection, there's a fetchcallback guard which uses
time(2) to prevent multiple fetches within a second, caching
the last result.  However, that means we still call time(2) on
each callback (per-metric) and in the failure case we call the
full refresh that often as well.

Added a per-fetch hook and moved the refresh logic there (with
the one-second caching guard still in place) to improve this a
little.