Further ease-of-use changes for supporting SSL connections.
If the per-user shared NSS database does not exist (e.g. via firefox,
pidgin, chrome, or some other application) we now create it on the
fly as a new-form sqlite3 NSS DB, including creating the path to it
(with appropriate directory modes) as needed.

Additionally, support code for the addition of certificates to user
databases by the monitor tools is here as well.  This can require
interaction with the user, which at this stage is done for console
tools only.  Longer term, will need a callback driven model so that
pmchart can play in this sandpit too.  However its certain that the
authentication changes will require similar treatment, so planning
to tackle those together to reduce API churn.