Make the server side nssdb access check more liberal
Previously we were only erroring out if the database does
not exist at all (ENOENT).  However, its also a possibility
that the permissions are such that the "pcp" user cannot
read it, as I came across in QA when not following the SSL
setup recipe to the letter (missed chmod, default is 0700).