Casting for character classification routines
We were being sloppy and inconsistent about casting the argument
to the character classificaton routines like isalpha(), isspace(),
isdigit(), isprint(), toupper(), tolower(), etc.

If the argument is type char, it needs to be cast to (int) to
be correct for all implementations of these library routines.

Without this we get compilation warnings on NetBSD, for
example.