Add a PCP_USER setting, replacing the hard-coded value
Serhei was building PCP using the configure-me-sideways
option to get a local developer build and found there's
still a need to become root, in order to use the start
scripts.  These want to change user to "pcp" nowadays,
but this may not exist in such a build environment.

This provides configure option, environment variable and
a pcp.conf setting allowing this to be changed.  This is
still achievable using the more manual -U option to each
of the daemons/agents as well.