Further refactoring of common server process (pmcd/pmproxy) code
Share a bit more code between pmcd and pmproxy, where ports are
setup and simplify the API a little in the process.  This makes
adding the missing -p option to pmproxy, trivial, and consistent
with pmcd handling.

Updated man pages a little for both pmcd and pmproxy in the port
specification area.

Small improvement in error handling around double-bind attempts,
which was picked up by a QA test - we were double-reporting this
due to the IPv6 bind attempt always being made (and reported),
its fine (better) to bail out on the first error as we have done
in the past.