Update QA test 273 to handle getaddrinfo looping over addresses
In changes recently made for IPv6 support we are iterating over
potentially a number of addresses, possibly both IPv4 and IPv6.
This causes the test to fail if the first address is not the one
used for the connection, as additional diagnostics are produced.
Handle this situation by "squashing" repeated occurrences of the
specific diagnostic in question into just one occurrence, which
matches the original output and intent.