Fix recent networking-related regression uncovered by test 131
With the introduction of getaddrinfo and looping over the set of
addresses it returns we were inadvertantly clobbering the return
codes and forcing ECONNREFUSED to always be returned.  Instead,
we now go with the return code of the last attempt, since it's
entirely plausible they are all failing for the same reason and
that there's handy diagnostic clues being thrown away.  This was
the case in qa test 131, which exercises the connection timeout
logic.

Nearly missed this one thanks to a series of unfortunate QA host
reconfigurations, esp with Mac OS X hosts sometimes going into a
deep sleep and not even responding to ssh logins anymore (ugh).