Fix further issues to get IPv4-only hosts working again
This code has been tested on a kernel booted with the kernel
flag ipv6.enabled cleared (bootup option), which picked up a
series of errors from coding assumptions.

The most problematic is the fact that the setsockopt call to
request a socket be made IPv6-only fails silently in kernels
booted this way.  This is the cause of the strange diagnostic
from a (later) bind call, as we will incorrectly attempt to
bind the same IPv4 port twice then.  This also resulted in a
cascading error (now fixed too):

Only add the listen file descriptor into the select mask once
we are certain its setup will succeed (previously we added it
before all possible error paths were evaluated).  This fixes
a sigsegv in the server code after the error paths have been
exercised.

Other small issues resolved in the process:
- Double-reporting of error message between OpenRequestPorts
and its helper OpenRequestSocket.
- For any setsockopts calls that are non-fatal, we now report
warning and not error level diagnostics.
- Improved error checking in NSS-enabled socket creation paths
which lacked an error check on __pmSetDataIPC (although the
__pmSetSocketIPC following was checking - should be the other
way around to correctly detect the malloc failure case).