Slackware build changes
Not slackware packaging yet (even though some changes in build/slack),
but with these changes we can build a tarball distribution for 64-bit
and 32-bit slcakware distros.

Note slackware may have rpm installed, but we do _not_ want to try
building packages on this distro with rpm, hence the configure.in
changes.