Reinstate fedora pcp.spec non-auto-generated RPM spec
The fedora spec here is a copy of the spec used in other
fedora build processes.  We don't really have opportunity
to generate it, match it up to build-configuration, etc,
during a PCP build.  Thats OK though, the potential paths
for config files are not varied with fedora, there can be
only the one.  We have to inline Kens magic move script,
for the above reasons.

Also update the fedora spec in preparation for the latest
PCP release.