Resolve issues in the RPM spec files around PCP_SYSCONF_DIR
The original strategy of reusing the packaging scripts from
deb packaging cannot work for RPM.  This is because the pre
script (where is resided) was moving files (as planned) but
they were later clobbered by the new files from the package
%files installation.

Revised to use the same basic strategy (examining the files
already installed) but only build up the needed commands in
the %pre section, then execute them in %post.  Its all very
nasty stuff but doesn't appear to be any other strategy for
keeping existing installations functional post-upgrade with
this kind of config file move (afaics, anyway).  Went for a
minimal fix at this stage of the release and have performed
extensive testing today (both Fedora and RHEL).

Updated CHANGELOGS etc in preparation for release now.