QA bigun PMDA - rework Install script
This is an odd PMDA as it is not installed in $PCP_VAR_DIR/pmdas and
so the installation logic is a little different.

Previously it only worked when run from the "testsuite" directory,
but I run all my QA out of the git tree qa directory ... these changes
allow it to work independent of where QA is run from.

The PMDA help text is a problem because this is a DSO PMDA, and this
is only available if Install is run from the "testsuite" directory
... this is not a problem for QA, so add a comment explaining this
in the PMDA source.