Correct an issue in the pmdasimple help text file
Found while looking into addition of the python PMDA module.

The indom numbering for the color and time indoms were not
zero-indexed as they should be, but rather were indexed from
one on.  This numbering is zero-based for both the C and perl
implementations.

QA test 716 now exercises this area, as well as adding further
checking to the descriptor handling in the perl implemenation.