Rework libpcp feature-test API internally for dynamic evaluation
With IPv6 becoming a runtime test (rather than build time), this
code really didn't support such extension.  Refactor it to use a
feature-detecting callback which allows dynamic feature testing.

As a special once-off, never-to-be-repeated bonus offer, removed
the two new statics in libpcp for IPv6 - not much gain to caching
these results forever, I think (especially for the one use case
which sets up ports for the servers - only executed at startup of
the daemons).