pmcd client connection tracing, qa/169
Fix for failing qa/169 which involved an unintended DNS timeout
in gethostbyaddr() at a critical point in the concurrent timeout
handling for both a failing PMDA and a client request.

In pmcd, do not do reverse DNS lookups for trace diagnostic,
report IP addr only.  Also rework this code to handle the mix
{ secure sockets enabled, secure sockets disabled } x { ipv4, ipv6 }.

Involves exposing the internals of pmcd's ClientInfo struct to
libpcp_pmcd, and getting some of the internal socket state from
libpcp visible in libpcp_pmcd.