pmwebapi: correct response memory-management snafu
In an earlier commit, I naively changed a static char[] buffer to a
plain auto char[] (on the stack) to construct a response message for
the /context? pmapi call, but neglected to change the
MHD_create_response_from_buffer description of that buffer.  It was
being reused, overwritten, corrupted.  Whoops.  Use ..._MUST_COPY.