libpcp - rework error path for __pmOpenLog
On the error path when the new log cannot be opened for writing,
the logic was not quite correct, and oldstream was not properly
restored ... this left applications like pmlogger writing stderr
into the void.

The problem is confined to the (rare, except for QA/083 and QA/524)
situation where the freopen() fails and we're able to recover
... but then did not leave the input oldstream attached to the
recovered i/o stream.

The fix is not without some issues ... refer to the comment in
the code.

Problem is actually in the internal routine logreopen() which is
called by both __pmOpenLog() and __pmRotateLog(), although the
likely use of the latter is unlikely to expose the problem.