qa/062 - remove ipv6 tests
This is the test that recently has been running for a _very_ long
time (like thousands of seconds) ... it now runs in under 2 secs.

For the src/chkacc?.c apps, add -4 (ipv4), -6 (ipv6) and -D cmd line
options ... default is -4.

Populate /etc/hosts with the IP addresses and bogus host names
needed to speed up the reverse DNS.

I'll leave it to someone who understands the ipv6 code and coverage
elsewhere in the QA suite to decide if and how it is appropriate
to reintroduce some variant of the ipv6 testing that used to be
in QA/062 in a (new) QA test.