libpcp/err.c - fix small initialization logic error
In pmErrStr_r() we have a one-trip initialization to optionally
find the "Unknown error ..." message buffer from libc and strdup()
it.  The code was not correctly dealing with a (possibly)
negative number after the "Unknown error " part of the message.

Found while running the error injection QA tests and we were not
seeing the fault in the strdup() here because the PM_FAULT_POINT
was not being executed as a result of the logic error.