QString toAscii decorruption
A recurrent coding pattern in the code aimed to get C "const char*"
pointers out of QString objects, with code that looks like this:

          (const char*) object.toAscii()

This is OK, using the QByteArray conversion-operator, as long as the
value is used before the QByteArray temporary is itself deleted.
However, in a few cases, the value is instead stored, or returned,
leading to corruption and deceit.  These have been reworked:

* src/chart/chartdialog.cpp (createChartPlot): strdup the nlabel.
* src/chart/colorscheme.cpp (colorSpec): Inline the rgbi string.
* src/chart/exportdialog.cpp (exportFile): Inline the strcmp args.
* src/chart/view.cpp (saveChart): Simplify and inline metric
  name/instance strings.
  (saveView): strdup incoming file name.
* src/libqmc/qmc_metric.h (*Ascii()): strdup return value.
* src/libqmc/qmc_source.h (*Ascii()): strdup return value.