Initial version of the MMV python wrapper module
ctypes wrapper for libpcp_mmv python client programs - not
yet fully functional, but not too far off either.

Exercised by a test_mmv python script, and test 704 which
is intended to be a full end-to-end test (all the way from
metric generation to verifying the metrics via pmdammv).

Refactored the duplicated _check_agent from 255 & 578 and
made use of it here in 704 to ensure sanity in the initial
testing environment.