Further work on the python MMV module interface
This is now functional, and metrics and instances can be created
from an instrumented python applications.  Hooray.

pylint still has one or two issues, but its giving a pretty high
score (~9.5) now too, so most of the cleanup work is complete.

In addition, found the MMV behaviour of allowing a non-running
process-flagged MMV app was not there in the C library (parfait
allows this for short-lived java apps, and its handy), so added
it in to use during QA.