Diagnostics - add DBG_TRACE_DESPERATE support
One level of additional verbosity control in the diagnostics.

Set from the command line with one or more flags, e.g -D pdu,desperate
and execution is guarded by a predictate like
    if ((pmDebug & DBG_TRACE_PDU) && (pmDebug & DBG_TRACE_DESPERATE))

The intention is that only the desperate developer would (a) add
diagnositics under the conjuncted DBG_TRACE_DESPERATE control,
(b) be interested in the output from these diagnostics.

For example uses in libpcp, see the subsequent commits to pdu.c and ipc.c.