libpcp - killer bug in PCP PDU layer exposed by secure sockets
One of the long-standing assumptions in the PCP PDU layer, is
that since we only ever write PDUs with a fixed sized header
plus a variable amount of data, on the receiver side an atomic
read()/recv() requesting the header size (or a whole word and
word-aligned part thereof) will return _all_ of the requested data,
else it is treated as an PM_ERR_IPC error and the show is over.

For regular sockets (pmcd/pmproxy/pmda/client connections) and files
(PCP archives) this assumption has stood the test of time across more
TCP/IP implementations than you could hit with a big stick.

But with secure sockets lurking under the covers and the various
and assorted handshake, encyption and compression games in play,
the assumption no longer stands and we have seen requests for 4,
8, 12 bytes returning data of length 1, 5, 7, ... bytes followed
by the remainder of the data in one or more subesquent recv()s
(so returning a shorter than expected buffer it is _not_ the end
of the data stream).

This commit re-engineers the _entire_ PDU assembly logic in
pduread() to match the brave new world reality.

With this change, secure socket connections from clients to pmcd
and/or pmproxy are working for me, and all of QA is passing on
at least one host so far.