Remove the code behind TIMEOUT_ASYNC (and back-compat macro)
The pduread code is obfuscated enough already without having
this in the way.  While reviering other changes in the area,
this was found wanting in several areas.  Nothing is using
it anymore, so just return an error if ASYNC requested.  In
a major release or two, we could remove the macros as well.