Fix the build when using older PCP headers (context flags).
Should have listened more closely to what Mark was telling me,
not that it *might* fail but rather that it *did* fail!?

Despite pre-emptive attempts to avoid this class of build issue
one of the affected source files did not include the version.h
header, hence did not pull in the back-compat goo.  Make It So.