Early stages of protocol level support for per-user authentication
This commit adds the necessary PDU encoding/decoding for SASL-based
authentication to libpcp.

The routines backing the pmconfig(1) utility are updated so that it
reports on the presence/absence of libpcp authentication support as
well.

The pmcd pmda is updated to export the PDU traffic in/out count for
the user authentication PDU, and an additional feature metric.  Also
tightened up the bounds checking in the fetch routine for pmdapmcd's
PDU counters.

Build system is updated to sniff out the presence of a SASL library
and header files, and incorporate their use into the libpcp build.

QA tests which make use of pmcd.pdu_{in,out} metrics are updated to
cater for the additional PDU (qa/375), and the pducrash.c utility
has been updated to verify handling of the PDU contents (qa/513).

Finally, pmcd now advertises the per-user authentication capability
in the feature bits it sends to clients - no clients notice it yet,
however (like a tree falling in a forest...).