Introduce $PCP_SECURE_DB_METHOD environment var
This is an attempt to overcome the problems associated with
hardcoding the sql: method into the certificate and key database
management, which fails badly on older platforms, e.g. CentOS 5.9
(and probably other, non-Linux platforms) where the sql: method
is not supported and using export PCP_SECURE_DB_METHOD='' seems
to restore old functionality for the non-secure socket users.

Note that $NSS_DEFAULT_DB_TYPE is not an option as this apparently
intended for compatibility support in the other direction,
as "sql" appears to be the only well-defined value for this
environment variable to enable the "new" method.