Experimental work toward extending the host specification APIs
I've been trying out some extensions to the host specification
API (which is the API used to parse the "-h <hostname>" option
in the client tools).  I'm seeking a way for all tools to pass
in information like whether to use secure/regular sockets, and
compression, and per-user authentication information, and also
any other future mechanism we might need.

This code implements support for an extended syntax, like:

    pcp://nas1.servers.com:44321@firewalls.r.us?compress
    pcps://nas1.servers.com?user=otto&pass=blotto&compress

It is backwards-compatible with the existing host syntax too.
It's not yet used there, but the plan is to make use of the
__pmParseHostAttrsSpec deep in pmNewContext (it designed as a
drop-in __pmParseHostSpec replacement).

Added __pmParseHostAttrsSpec(3), updated __pmParseHostSpec(3),
and wrote some new QA tests (719, 720).  As a by-product, I've
found __pmUnparseHostSpec to be almost completely unuseable in
all but the simplest scenarios, with no error checking and with
a strange calling convention.  As this is an almost-never-used
double-underscore routine, I've fixed it, tested and documented
the new behaviour.