Rename python test script to match others and ensure build passiness
The py extension is magic to certain package managers, and if used in
a script name causes follow-on effects (such as byte-compilation) and
at the moment, we don't cater for this.
error: Installed (but unpackaged) file(s) found:
   /var/lib/pcp/testsuite/660.pyc
   /var/lib/pcp/testsuite/660.pyo

For now, use the same approach as the other python scripts to avoid
this pitfall.