configury: enable gnumake parallel (-jN) builds
With the explicit recursion logic for GNUmakefile SUBDIRS, it was
impossible for a top-level "make -jN" to get anything done, since the
jobserver-related options were not being propagated.  Fixed this by
adding a strategic "+" punctuation to the recursion rule.  Added
".NOTPARALLEL:" to several sub-GNUmakefiles that suffer from race
conditions or insufficient dependency specifications.

See also [1] wherein the same problem is seen, due to the same
provenance of XFS & PCP makefile/configury.

[1] http://blog.flameeyes.eu/2008/11/for-a-parallel-world-case-study-n-4-jobserver-unavailable