Create a single directory housing links to critical libraries
This commit adds population of a lib directory into the build,
initially to simplify building of the scripting language bits.
There are other components that can be simplified using this
too, qa/src springs to mind.

In the process, this fixes the issue Dave Brolley is seeing on
a build machine with no PCP libraries in the root, as a result
of recent python extensions.