Change to start pmwebd up as a regular daemon
Init scripts, logfiles, non-privileged user options,
configuration files and packaging for all platforms;
all added in for pmwebd.  Alot based on the pmproxy
handling of running itself as a daemon, and one/two
cosmetic updates to the pmproxy processes are made
as a result.

Man page updates to document the new options, code
and associated QA test updates too of course.

Changed the '-C' option to '-c' in anticipation of
using -C down the track consistently with pmcd and
pmproxy.

Makefile changes in pmwebapi to be more consistent
with the way other PCP daemons are built, and also
to ensure source tarball builds will have all code
even when libmicrohttpd is not installed within the
build environment.

Added the standard libpcp debugging option -D into
pmwebd as well, it will surely be needed someday.