Changes to pmdagfs2 to add addtional filesystem metrics and their corresponding help text.
Changed the location on where the indom cache fetched filesystem
names to /sys/fs/gfs2/ to allow the additional query for the device
id of the filesystem which is needed for the gfs2_glock_lock_time
tracepoint metrics. Also added the required code to support the
additional metrics.

Added metrics for glstats "/sys/kernel/debug/gfs2/<fs>/glstats" these
metrics can count the number of glocks that exists currently for the
filesystem and separate how many per type of glock.

Added metrics for the gfs2_glock_lock_time tracepoint. These metrics
determine the current *worst* glock for the filesystem at the time of
query. The data is fetched from "/sys/kernel/debug/tracing/trace_pipe"
which lists the values of glocks for all filesystems as they are updated
in (real-time). We read through the pipe taking all of the buffered values
and calculate which is our worst glock, the device id for each filesystem
is used to separate the locks for each filesystem.

Created some additional qa tests in "qa/654".